backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
httpx==0.28.0
langchain==0.3.14
langchain-openai==0.2.14
langchain-google-genai==2.0.10
langchain-anthropic==0.3.0
langchain-community==0.3.14
langsmith==0.2.8
upstash-redis==1.1.0
spacy==3.8.3
numpy==1.26.4
scipy==1.14.1
pandas==2.2.3
python-multipart==0.0.12
pytest==8.3.4
pytest-asyncio==0.24.0